refactor: split api/routes.py into focused route modules
